+++ src/pages/games/AppleGame/game/model/appleGame/apple.ts
