+++ src/pages/games/AppleGame/game/model/appleGame/goldenApple.ts
