+++ src/pages/games/AppleGame/game/model/appleGame/plainApple.ts
