+++ src/pages/games/AppleGame/game/model/common/click.ts
