+++ src/pages/games/AppleGame/game/model/common/drag.ts
