+++ src/pages/games/AppleGame/game/model/common/particle.ts
