+++ src/pages/games/AppleGame/game/model/snackGame/goldSnack.ts
