+++ src/pages/games/AppleGame/game/model/snackGame/plainSnack.ts
